requirements.txt
django==5.2.7
djangorestframework
django-cors-headers
python-dotenv
pillow
twilio
pyotp
pandas
openpyxl
django-allauth
dj-rest-auth
cryptography
